+++ jgclark.Dashboard/src/constants.js
@@ -1,7 +1,7 @@
 // @flow
 //-----------------------------------------------------------------------------
 // Constants for Dashboard code
-// Last updated 2025-02-21 for v2.2.0, @jgclark
+// Last updated 2025-02-22 for v2.2.0, @jgclark
 //-----------------------------------------------------------------------------
 import pluginJson from '../plugin.json'
 import type { TSectionDetails, TSectionCode } from './types'
@@ -33,7 +33,7 @@ export const allSectionCodes: Array<TSectionCode> = allSectionDetails.map((s) =>
 
 export const allCalendarSectionCodes = ['TB', 'DT', 'DY', 'DO', 'LW', 'W', 'M', 'Q']
 
-export const sectionDisplayOrder = ['SEARCH', 'TB', 'DT', 'DY', 'DO', 'LW', 'W', 'M', 'Q', 'TAG', 'OVERDUE', 'PRIORITY', 'PROJ']
+export const sectionDisplayOrder = ['QSEARCH', 'SEARCH', 'TB', 'DT', 'DY', 'DO', 'LW', 'W', 'M', 'Q', 'TAG', 'OVERDUE', 'PRIORITY', 'PROJ']
 
 // change this order to change which duplicate gets kept - the first on the list. Should not include 'dontDedupeSectionCodes' below.
 export const sectionPriority = ['TB', 'TAG', 'DT', 'DY', 'DO', 'W', 'M', 'Q', 'PRIORITY', 'OVERDUE']
@@ -41,4 +41,4 @@ export const sectionPriority = ['TB', 'TAG', 'DT', 'DY', 'DO', 'W', 'M', 'Q', 'P
 // Those sections we can't or shouldn't attempt to dedupe:
 // - TB as its for info only
 // - PROJ as it isn't about paragraphs, but notes
-export const dontDedupeSectionCodes = ['PROJ', 'SEARCH']
+export const dontDedupeSectionCodes = ['PROJ', 'QSEARCH', 'SEARCH']