+++ jgclark.Dashboard/CHANGELOG.md
@@ -1,6 +1,19 @@
 # What's changed in 🎛 Dashboard plugin?
 For more details see the [plugin's documentation](https://github.com/NotePlan/plugins/tree/main/jgclark.Dashboard/).
 
+- TODO: Search: Add in prevent tasks under heading (from Perspective)
+- TODO: UI to set the searchTerm and search options.
+- TODO: Search: Update background colour slightly
+- TODO: Prevent banner warning when completing non-standard scheduled items (that don't have a `<date` component)
+- TODO: Allow ignoring future dated items found in regular and calendar notes as well.
+
+## [2.2.0.a2] 2025-02-22
+- basic Search Bar in Header is working
+- Improved some icons and their colouring
+- Search section will not sync items even if set to do so in SearchExtensions.
+- Fixed ordering of Search section
+- Added setting to turn off future tasks. _Currently only works on items in future calendar notes._
+
 ## [2.2.0.a1] 2025-02-21
 ### New
 - 'Search' section started, and integrated in all relevant search and display logic.
@@ -11,11 +24,8 @@ For more details see the [plugin's documentation](https://github.com/NotePlan/pl
   - ISO end date for calendar notes (optional, default is empty)
 - Can close section when finished with
 - WIP: Start to make Search Bar in Header.
-- TODO: Add in prevent tasks under heading (from Perspective)
-- TODO: UI to set the searchTerm and search options.
 ### Changed
 - improves display of non-standard scheduled items (that don't have a `<date` component)
-- TODO: prevent banner warning when completing non-standard scheduled items (that don't have a `<date` component)
 ### Fixed
 
 ## [2.1.10] 2025-02-16